+++ src/main.ts
@@ -3,11 +3,13 @@ import "./style.css";
 import Toast from "vue-toastification";
 import "vue-toastification/dist/index.css";
 import App from "./App.vue";
+import router from "./routes/index";
 
-createApp(App)
-  .use(Toast, {
-    transition: "Vue-Toastification__bounce",
-    maxToasts: 5,
-    newestOnTop: true,
-  })
-  .mount("#app");
+const app = createApp(App);
+app.use(router);
+app.use(Toast, {
+  transition: "Vue-Toastification__bounce",
+  maxToasts: 5,
+  newestOnTop: true,
+});
+app.mount("#app");
