+++ src/pages/about.vue
@@ -1,0 +1,6 @@
+<template lang="">
+  <div>
+    <h1>About vie</h1>
+  </div>
+</template>
+<script setup></script>
